requirements.txt
blinker==1.9.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
et_xmlfile==2.0.0
Flask==3.1.2
fonttools==4.60.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.9
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.6
mdurl==0.1.2
numpy==2.3.3
openpyxl==3.1.5
packaging==25.0
pandas==2.3.2
pillow==11.3.0
Pygments==2.19.2
pyparsing==3.2.4
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.5
rich==14.1.0
six==1.17.0
sqlparse==0.5.3
tabulate==0.9.0
tzdata==2025.2
urllib3==2.5.0
Werkzeug==3.1.3
